Fixes enabling successful builds on linux.
Fixes to include files that windows.h provides
on windows.

Fixes to include files required on linux platform.

Fixes to linker options in CMake.

+++ src/game/memory/newdel.cpp
@@ -16,12 +16,12 @@
 #include "newdel.h"
 #include "alloc.h"
 
-void* operator new(unsigned int size)
+void* operator new(size_t size)
 {
     return Alloc(size, MEM_NEW);
 }
 
-void* operator new[](unsigned int size)
+void* operator new[](size_t size)
 {
     return Alloc(size, MEM_NEW);
 }
